third_party/WebKit/LayoutTests/external/wpt/beacon/headers/header-referrer-strict-origin.https-expected.txt

This is a testharness.js-based test.
FAIL Test referer header https://web-platform.test:8444/beacon/resources/ assert_equals: Correct referrer header result expected "https://web-platform.test:8444/" but got "https://web-platform.test:8444/beacon/headers/header-referrer-strict-origin.https.html"
FAIL Test referer header http://web-platform.test:8001/beacon/resources/ assert_true: SendBeacon Succeeded expected true got false
Harness: the test ran to completion.
